Make sub_product skus searchable
Currently Producthelper->getObject only returns attribute values from child
products if the parent product attribute value is null. We would like to be able to
search for the sku of both parent and child products. This patch adds an
exception to getObject to include all the skus in the returned object.